+++ GeecsBluesky/geecs_bluesky/utils.py
@@ -72,6 +72,7 @@ RESERVED_DEVICE_ATTRIBUTES: frozenset[str] = frozenset(
         "mark_abandoned",
         "missed_shot",
         "name",
+        "native_image_save",
         "parent",
         "plugin_backed",
         "prepare",